application/external_apps/simplechat-mcp-server/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
Authlib==1.6.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
cyclopts==3.22.5
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastmcp==2.10.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mcp==1.12.3
mdurl==0.1.2
msal==1.33.0
openapi-pydantic==0.5.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.26.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
